+++ CHANGELOG.md
@@ -158,3 +158,4 @@
 # 0.2.4
 * Remove `Pandora.Core.Transformation` module and move `~>` to `Functor` module
 * Define `Adaptable` instances for comonad transformer schemes
+* Make `position` and `access` methods of `Store` adaptable